Flask-project/requirements.txt
# Installing some dependencies used for the project
flask
requests
google-generativeai
